Upgrade to hibernate 3.2.7.ga and Spring 3.2.18.RELEASE.
pom.xml:
- Upgrade hibernate from 3.1.3 to 3.2.7.ga.
- Upgrade h2 database from 1.3.172 to 1.4.199
- Upgrade Spring from 2.5.6 to 3.2.18.RELEASE.
- Add antlr:antlr:2.7.7 to supersede hibernate's 2.7.6 version.

EnvironmentFactory.java:
- 'environment' is a specific bean name in Spring 3.2, so this particular
  bean has been renamed to 'environmentWCT'.

Role.java, RolePrivilege.java, User.java:
- Hibernate now requires that JOIN or INNER JOIN be used to chain through
  one-to-many relationships.

AbstractGroupTypeView.java:
- Removed foreign-key for many-to-one relationships. Foreign keys are not
  allowed on table views.

TargetInstanceDAOImpl.java:
- Hibernate seems to now provide automatic conversion to Long for select counts
  so no need to do casts in countQueueLengthForTarget.

WaybackIndexer.java:
- Tracks the last file not indexed, which can be useful for debugging purposes
  when the indexing does not complete (info log message lists the file).

Additional notes:
- When running the Jetty server, Jetty will print out a list of classes that
  appear in more that one jar file. The following conflicts will exist:
  - xmlpull:xmlpull and xpp3:xpp3_min - both brought via drools xstream
    transitive dependency, no way to remove the conflict.
  - jdt core and jdt ecj - jdt core comes in via Jetty, no way to remove the
    conflict
  - org.apache.taglibs:taglibs-standard-impl and javax.servlet.jstl:jstl
    - jstl is required for the fmt tag, which isn't included in taglibs-standard.
      Jetty is deploying taglibs-standard-impl, no way to remove the conflict.
- Hibernate will still generate errors when setting up views (specifially the
  ABSTRACT_TARGET_GROUPTYPE_VIEW). This doesn't seem to stop the application
  from working. This issue may be resolved when Hibernate is further upgraded
  and JPA notation is used instead of the generated .hbm.xml files.

+++ src/main/java/org/webcurator/domain/TargetInstanceDAOImpl.java
@@ -678,7 +678,7 @@ public class TargetInstanceDAOImpl extends HibernateDaoSupport implements Target
 					query.setDate("ed", new Date());
 					query.setLong("toid", targetOid);
 					
-					return ((Integer)query.list().get(0)).longValue();
+					return query.list().get(0);
 				}
 			}
 		);
